. encdec.h, encdec.cpp, encdecg.cpp, sinit.cpp, sinit/Makefile.am: moved sinit directory from security/sinit to monitor/sinit, and split sinit to generate a shared library, so that enc/dec functions can be shared with the security module and monitor modules. . configure.ac, rpcf.spec.in, montor/Makefile.am, appmon/Makefile.am, SimpleAuthsvr.cpp, saproxy.cpp, security/Makefile.am: moved the scattered enc/dec function to the encdec.cpp and encdecg.cpp, and link to the shared library librpcfenc.so instead.
