+++ bldpkg/rpcf.spec.in
@@ -97,6 +97,7 @@ make install DESTDIR=$RPM_BUILD_ROOT
 %{_libdir}/liblogcli.so*
 %{_libdir}/libregfs.so*
 %{_libdir}/libappmancli.so*
+%{_libdir}/librpcfenc.so*
 %{_mandir}/man1/ridlc.1.gz
 %{_mandir}/man1/rpcrouter.1.gz
 %{_mandir}/man7/ridl.7.gz
