+++ monitor/Makefile.am
@@ -1,4 +1,4 @@
-SUBDIRS=registry appmon usermgmt client
+SUBDIRS=registry sinit appmon usermgmt client
 if ENABLE_JS
 SUBDIRS += oinit
 endif
